refactor claim pages, adding wallet recognition and loading stage
